the crud is working

+++ src/main/java/com/gms/gym/repository/WorkoutRepository.java
@@ -1,8 +1,7 @@
 package com.gms.gym.repository;
 
-
 import com.gms.gym.entity.Workout;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkoutRepository extends JpaRepository<Workout, Integer> {
+public interface WorkoutRepository extends JpaRepository<Workout, Long> {
 }
